mm, treewide: make addr parameter of free_pages() void *
Most callers of free_pages() and free_page() cast a pointer type to
unsigned long.

Make addr parameter of free_pages() and free_page() 'void *' and update
call sites with the following semantic patch and a bit of manual tweaking.

    @@
    identifier addr;
    expression order;
    @@
    (
    - free_pages(addr, order)
    + free_pages((void *)addr, order)
    |
    - free_page(addr)
    + free_page((void *)addr)
    )

    @@
    type T != void *;
    expression addr, order;
    @@
    (
    - free_pages((T)addr, order)
    + free_pages(addr, order)
    |
    - free_page((T)addr)
    + free_page(addr)
    )

+++ arch/powerpc/include/asm/book3s/64/pgalloc.h
@@ -39,9 +39,9 @@ static inline pgd_t *radix__pgd_alloc(struct mm_struct *mm)
 static inline void radix__pgd_free(struct mm_struct *mm, pgd_t *pgd)
 {
 #ifdef CONFIG_PPC_64K_PAGES
-	free_page((unsigned long)pgd);
+	free_page(pgd);
 #else
-	free_pages((unsigned long)pgd, 4);
+	free_pages(pgd, 4);
 #endif
 }
