+++ arch/x86/kvm/vmx/nested.c
@@ -7441,7 +7441,7 @@ void nested_vmx_hardware_unsetup(void)
 
 	if (enable_shadow_vmcs) {
 		for (i = 0; i < VMX_BITMAP_NR; i++)
-			free_page((unsigned long)vmx_bitmap[i]);
+			free_page(vmx_bitmap[i]);
 	}
 }
