+++ arch/x86/xen/mmu_pv.c
@@ -151,7 +151,7 @@ static int alloc_discontig_frames(unsigned int order)
 
 	spin_unlock_irqrestore(&xen_reservation_lock, flags);
 
-	free_pages((unsigned long)old_array, old_order - MIN_CONTIG_ORDER);
+	free_pages(old_array, old_order - MIN_CONTIG_ORDER);
 
 	return 0;
 }
@@ -1479,7 +1479,7 @@ static void xen_pgd_free(struct mm_struct *mm, pgd_t *pgd)
 	pgd_t *user_pgd = xen_get_user_pgd(pgd);
 
 	if (user_pgd)
-		free_page((unsigned long)user_pgd);
+		free_page(user_pgd);
 }
 
 /*
