+++ drivers/media/pci/intel/ipu6/ipu6-mmu.c
@@ -158,7 +158,7 @@ static int get_dummy_page(struct ipu6_mmu_info *mmu_info)
 	return 0;
 
 err_free_page:
-	free_page((unsigned long)pt);
+	free_page(pt);
 	return -ENOMEM;
 }
 
@@ -167,7 +167,7 @@ static void free_dummy_page(struct ipu6_mmu_info *mmu_info)
 	dma_unmap_single(mmu_info->dev,
 			 TBL_PHYS_ADDR(mmu_info->dummy_page_pteval),
 			 PAGE_SIZE, DMA_BIDIRECTIONAL);
-	free_page((unsigned long)mmu_info->dummy_page);
+	free_page(mmu_info->dummy_page);
 }
 
 static int alloc_dummy_l2_pt(struct ipu6_mmu_info *mmu_info)
@@ -196,7 +196,7 @@ static int alloc_dummy_l2_pt(struct ipu6_mmu_info *mmu_info)
 	return 0;
 
 err_free_page:
-	free_page((unsigned long)pt);
+	free_page(pt);
 	return -ENOMEM;
 }
 
@@ -205,7 +205,7 @@ static void free_dummy_l2_pt(struct ipu6_mmu_info *mmu_info)
 	dma_unmap_single(mmu_info->dev,
 			 TBL_PHYS_ADDR(mmu_info->dummy_l2_pteval),
 			 PAGE_SIZE, DMA_BIDIRECTIONAL);
-	free_page((unsigned long)mmu_info->dummy_l2_pt);
+	free_page(mmu_info->dummy_l2_pt);
 }
 
 static u32 *alloc_l1_pt(struct ipu6_mmu_info *mmu_info)
@@ -234,7 +234,7 @@ static u32 *alloc_l1_pt(struct ipu6_mmu_info *mmu_info)
 	return pt;
 
 err_free_page:
-	free_page((unsigned long)pt);
+	free_page(pt);
 	return NULL;
 }
 
@@ -340,7 +340,7 @@ static int l2_map(struct ipu6_mmu_info *mmu_info, unsigned long iova,
 			dma = map_single(mmu_info, l2_virt);
 			if (!dma) {
 				dev_err(dev, "Failed to map l2pt page\n");
-				free_page((unsigned long)l2_virt);
+				free_page(l2_virt);
 				err = -EINVAL;
 				goto error;
 			}
@@ -732,7 +732,7 @@ static void ipu6_mmu_destroy(struct ipu6_mmu *mmu)
 			dma_unmap_single(mmu_info->dev,
 					 TBL_PHYS_ADDR(mmu_info->l1_pt[l1_idx]),
 					 PAGE_SIZE, DMA_BIDIRECTIONAL);
-			free_page((unsigned long)mmu_info->l2_pts[l1_idx]);
+			free_page(mmu_info->l2_pts[l1_idx]);
 		}
 	}
 
@@ -740,8 +740,8 @@ static void ipu6_mmu_destroy(struct ipu6_mmu *mmu)
 	free_dummy_page(mmu_info);
 	dma_unmap_single(mmu_info->dev, TBL_PHYS_ADDR(mmu_info->l1_pt_dma),
 			 PAGE_SIZE, DMA_BIDIRECTIONAL);
-	free_page((unsigned long)mmu_info->dummy_l2_pt);
-	free_page((unsigned long)mmu_info->l1_pt);
+	free_page(mmu_info->dummy_l2_pt);
+	free_page(mmu_info->l1_pt);
 	kfree(mmu_info);
 }
