+++ drivers/net/ethernet/mellanox/mlx5/core/lib/clock.c
@@ -1385,7 +1385,7 @@ static void mlx5_destroy_clock_dev(struct mlx5_core_dev *mdev)
 	}
 
 	if (mdev->clock_info) {
-		free_page((unsigned long)mdev->clock_info);
+		free_page(mdev->clock_info);
 		mdev->clock_info = NULL;
 	}
