+++ drivers/net/rionet.c
@@ -526,7 +526,7 @@ static int rionet_setup_netdev(struct rio_mport *mport, struct net_device *ndev)
 
 	rc = register_netdev(ndev);
 	if (rc != 0) {
-		free_pages((unsigned long)nets[mport->id].active,
+		free_pages(nets[mport->id].active,
 			   get_order(rionet_active_bytes));
 		goto out;
 	}
@@ -680,9 +680,8 @@ static void rionet_remove_mport(struct device *dev)
 		netif_stop_queue(ndev);
 		unregister_netdev(ndev);
 
-		free_pages((unsigned long)nets[id].active,
-			   get_order(sizeof(void *) *
-			   RIO_MAX_ROUTE_ENTRIES(mport->sys_size)));
+		free_pages(nets[id].active,
+			   get_order(sizeof(void *) * RIO_MAX_ROUTE_ENTRIES(mport->sys_size)));
 		nets[id].active = NULL;
 		free_netdev(ndev);
 		nets[id].ndev = NULL;
