+++ drivers/s390/char/sclp_config.c
@@ -123,7 +123,7 @@ static int sclp_ofb_send_req(char *ev_data, size_t len)
 		rc = -EIO;
 	}
 out:
-	free_page((unsigned long)sccb);
+	free_page(sccb);
 	return rc;
 }
