+++ drivers/s390/char/zcore.c
@@ -238,7 +238,7 @@ static int __init zcore_reipl_init(void)
 	if (rc || (__force u32)csum_partial(zcore_ipl_block, zcore_ipl_block->hdr.len, 0) !=
 	    ipib_info.checksum) {
 		TRACE("Checksum does not match\n");
-		free_page((unsigned long) zcore_ipl_block);
+		free_page(zcore_ipl_block);
 		zcore_ipl_block = NULL;
 	}
 	/*
@@ -269,7 +269,7 @@ static int __init zcore_reipl_init(void)
 			os_info_flags = 0;
 	}
 out:
-	free_page((unsigned long)os_info);
+	free_page(os_info);
 	return 0;
 }
