+++ net/sctp/protocol.c
@@ -1674,9 +1674,8 @@ err_register_defaults:
 	sctp_v4_pf_exit();
 	sctp_v6_pf_exit();
 	sctp_sysctl_unregister();
-	free_pages((unsigned long)sctp_port_hashtable,
-		   get_order(sctp_port_hashsize *
-			     sizeof(struct sctp_bind_hashbucket)));
+	free_pages(sctp_port_hashtable,
+		   get_order(sctp_port_hashsize * sizeof(struct sctp_bind_hashbucket)));
 err_bhash_alloc:
 	sctp_transport_hashtable_destroy();
 err_thash_alloc:
@@ -1715,9 +1714,8 @@ static __exit void sctp_exit(void)
 
 	sctp_sysctl_unregister();
 
-	free_pages((unsigned long)sctp_port_hashtable,
-		   get_order(sctp_port_hashsize *
-			     sizeof(struct sctp_bind_hashbucket)));
+	free_pages(sctp_port_hashtable,
+		   get_order(sctp_port_hashsize * sizeof(struct sctp_bind_hashbucket)));
 	kfree(sctp_ep_hashtable);
 	sctp_transport_hashtable_destroy();
